+++ NEWS.md
@@ -1,3 +1,10 @@
+# Qploidy2 1.17.0
+
+* Change `plot_cn_track` labels from `Est overall ploidy` to `Mean total depth/Z CN` and y-axis label `Copy Number` to `HMM-based CN`
+* Add argument `threshold.missing.samples` to `standardize`. It filters samples with fraction of markers containig missing information higher than
+defined by the argument. Default set to `1` (filter not applied).
+* Function `export_mappoly` to produce input file for `filter_aneuploidy` function in `MAPpoly`
+
 # Qploidy2 1.16.2
 
 * Add different VCF required fields for when `qploidy_read_vcf` is `geno = TRUE` or `FALSE`. When it is `FALSE`, `GT` is not required